Move CAA code into separate file in VA.
This is simply a code-move, nothing has been edited.
