+++ va/va_test.go
@@ -9,7 +9,6 @@ import (
 	"crypto/x509/pkix"
 	"encoding/base64"
 	"encoding/hex"
-	"errors"
 	"fmt"
 	"math/big"
 	mrand "math/rand"
@@ -28,7 +27,6 @@ import (
 
 	"github.com/golang/mock/gomock"
 	"github.com/jmhodges/clock"
-	"github.com/miekg/dns"
 	"golang.org/x/net/context"
 	"gopkg.in/square/go-jose.v2"
 
@@ -781,84 +779,6 @@ func TestValidateTLSSNI01NotSane(t *testing.T) {
 	test.AssertEquals(t, prob.Type, probs.MalformedProblem)
 }
 
-func TestCAATimeout(t *testing.T) {
-	va, _ := setup(nil, 0)
-	err := va.checkCAA(ctx, core.AcmeIdentifier{Type: core.IdentifierDNS, Value: "caa-timeout.com"})
-	if err.Type != probs.ConnectionProblem {
-		t.Errorf("Expected timeout error type %s, got %s", probs.ConnectionProblem, err.Type)
-	}
-	expected := "DNS problem: query timed out looking up CAA for always.timeout"
-	if err.Detail != expected {
-		t.Errorf("checkCAA: got %#v, expected %#v", err.Detail, expected)
-	}
-}
-
-func TestCAAChecking(t *testing.T) {
-	type CAATest struct {
-		Domain  string
-		Present bool
-		Valid   bool
-	}
-	tests := []CAATest{
-		// Reserved
-		{"reserved.com", true, false},
-		// Critical
-		{"critical.com", true, false},
-		{"nx.critical.com", true, false},
-		// Good (absent)
-		{"absent.com", false, true},
-		{"example.co.uk", false, true},
-		// Good (present)
-		{"present.com", true, true},
-		{"present.servfail.com", true, true},
-		// Good (multiple critical, one matching)
-		{"multi-crit-present.com", true, true},
-		// Bad (unknown critical)
-		{"unknown-critical.com", true, false},
-		{"unknown-critical2.com", true, false},
-		// Good (unknown noncritical, no issue/issuewild records)
-		{"unknown-noncritical.com", true, true},
-		// Good (issue record with unknown parameters)
-		{"present-with-parameter.com", true, true},
-		// Bad (unsatisfiable issue record)
-		{"unsatisfiable.com", true, false},
-	}
-
-	va, _ := setup(nil, 0)
-	for _, caaTest := range tests {
-		present, valid, err := va.checkCAARecords(ctx, core.AcmeIdentifier{Type: "dns", Value: caaTest.Domain})
-		if err != nil {
-			t.Errorf("checkCAARecords error for %s: %s", caaTest.Domain, err)
-		}
-		if present != caaTest.Present {
-			t.Errorf("checkCAARecords presence mismatch for %s: got %t expected %t", caaTest.Domain, present, caaTest.Present)
-		}
-		if valid != caaTest.Valid {
-			t.Errorf("checkCAARecords validity mismatch for %s: got %t expected %t", caaTest.Domain, valid, caaTest.Valid)
-		}
-	}
-
-	present, valid, err := va.checkCAARecords(ctx, core.AcmeIdentifier{Type: "dns", Value: "servfail.com"})
-	test.AssertError(t, err, "servfail.com")
-	test.Assert(t, !present, "Present should be false")
-	test.Assert(t, !valid, "Valid should be false")
-
-	_, _, err = va.checkCAARecords(ctx, core.AcmeIdentifier{Type: "dns", Value: "servfail.com"})
-	if err == nil {
-		t.Errorf("Should have returned error on CAA lookup, but did not: %s", "servfail.com")
-	}
-
-	present, valid, err = va.checkCAARecords(ctx, core.AcmeIdentifier{Type: "dns", Value: "servfail.present.com"})
-	test.AssertError(t, err, "servfail.present.com")
-	test.Assert(t, !present, "Present should be false")
-	test.Assert(t, !valid, "Valid should be false")
-
-	_, _, err = va.checkCAARecords(ctx, core.AcmeIdentifier{Type: "dns", Value: "servfail.present.com"})
-	if err == nil {
-		t.Errorf("Should have returned error on CAA lookup, but did not: %s", "servfail.present.com")
-	}
-}
-
 func TestPerformValidationInvalid(t *testing.T) {
 	va, _ := setup(nil, 0)
 
@@ -1035,18 +955,6 @@ func TestDNSValidationNoAuthorityOK(t *testing.T) {
 	test.Assert(t, prob == nil, "Should be valid.")
 }
 
-func TestCAAFailure(t *testing.T) {
-	chall := createChallenge(core.ChallengeTypeTLSSNI01)
-	hs := tlssni01Srv(t, chall)
-	defer hs.Close()
-
-	va, _ := setup(hs, 0)
-
-	ident.Value = "reserved.com"
-	_, prob := va.validateChallengeAndCAA(ctx, ident, chall)
-	test.AssertEquals(t, prob.Type, probs.ConnectionProblem)
-}
-
 func TestLimitedReader(t *testing.T) {
 	chall := core.HTTPChallenge01()
 	setChallengeToken(&chall, core.NewToken())
@@ -1087,24 +995,6 @@ func setup(srv *httptest.Server, maxRemoteFailures int) (*ValidationAuthorityImp
 		clock.Default(),
 		logger)
 	return va, logger
-}
-
-func TestParseResults(t *testing.T) {
-	r := []caaResult{}
-	s, err := parseResults(r)
-	test.Assert(t, s == nil, "set is not nil")
-	test.Assert(t, err == nil, "error is not nil")
-	test.AssertNotError(t, err, "no error should be returned")
-	r = []caaResult{{nil, errors.New("")}, {[]*dns.CAA{{Value: "test"}}, nil}}
-	s, err = parseResults(r)
-	test.Assert(t, s == nil, "set is not nil")
-	test.AssertEquals(t, err.Error(), "")
-	expected := dns.CAA{Value: "other-test"}
-	r = []caaResult{{[]*dns.CAA{&expected}, nil}, {[]*dns.CAA{{Value: "test"}}, nil}}
-	s, err = parseResults(r)
-	test.AssertEquals(t, len(s.Unknown), 1)
-	test.Assert(t, s.Unknown[0] == &expected, "Incorrect record returned")
-	test.AssertNotError(t, err, "no error should be returned")
 }
 
 func TestAvailableAddresses(t *testing.T) {
